major refactor of google analytics child process and bundle code
