+++ packages/truffle-core/lib/commands/init.js
@@ -18,10 +18,9 @@ var command = {
     var OS = require("os");
     var UnboxCommand = require("./unbox");
     const path = require('path');    
-    const cp = require('child_process');
-    const child = cp.fork(path.join(__dirname,  "../services/google-analytics"));
+    const googleAnalytics = require("../services/analytics");
 
-    child.send({ec: "initialization", ea: "truffle init"});
+    googleAnalytics.send({ec: "initialization", ea: "truffle init"});
 
     var config = Config.default().with({
       logger: console
