+++ packages/truffle-core/lib/commands/unbox.js
@@ -70,10 +70,8 @@ var command = {
     const Config = require("truffle-config");
     const Box = require("truffle-box");
     const OS = require("os");
-    const googleAnalytics = require('../services/google-analytics');
+    const googleAnalytics = require("../services/analytics");
     const path = require('path');
-    const cp = require('child_process');
-    const child = cp.fork(path.join(__dirname,  "../services/google-analytics"));
 
     const config = Config.default().with({
       logger: console
@@ -102,7 +100,7 @@ var command = {
         if (boxConfig.epilogue) {
           config.logger.log(boxConfig.epilogue.replace("\n", OS.EOL));
         }
-        child.send({ec: "initialization", ea: "truffle unbox"});
+        googleAnalytics.send({ec: "initialization", ea: "truffle unbox"});
         
         done(); 
       })
