+++ packages/truffle-core/lib/services/analytics/index.js
@@ -1,0 +1,18 @@
+require("./main.js");
+const analytics = {
+	send: function(eventObject) {
+		let analyticsPath;
+		const path = require('path');
+		if (typeof BUNDLE_ANALYTICS_FILENAME != "undefined") {
+		  analyticsPath = path.join(__dirname, BUNDLE_ANALYTICS_FILENAME);
+		} else {
+		  analyticsPath = path.join(__dirname, "./lib/services/analytics/main.js");
+		}
+
+		const cp = require("child_process");
+		const child = cp.fork(analyticsPath);
+		child.send(eventObject);
+	},
+}
+
+module.exports = analytics;
