+++ packages/truffle-core/lib/services/analytics/main.js
@@ -1,0 +1,9 @@
+const googleAnalytics = require("./google.js");
+
+process.on('message', function(eventObject) {
+	googleAnalytics.sendAnalyticsEvent(eventObject);
+
+	setTimeout(function(){
+		process.exit(0);
+	}, 5000);
+});
