+++ packages/truffle/cli.webpack.config.js
@@ -15,7 +15,7 @@ module.exports = {
   entry: {
     cli: path.join(__dirname, "../..", "node_modules", "truffle-core", "cli.js"),
     chain: path.join(__dirname, "../..", "node_modules", "truffle-core", "chain.js"),
-    analytics: path.join(__dirname, "../..", "node_modules", "truffle-core", "lib", "services", "google-analytics.js")
+    analytics: path.join(__dirname, "../..", "node_modules", "truffle-core", "lib", "services", "analytics", "index.js")
   },
   target: 'node',
   node: {
